refactor: move collect_imports_from_module to import_deduplicator
Move the collect_imports_from_module method from bundler.rs to
import_deduplicator.rs as a free function. This improves code organization
by placing the function alongside other stdlib import handling functions
like is_hoisted_import, add_stdlib_import, and add_hoisted_imports.

Changes:
- Convert method to free function with bundler as first parameter
- Update imports in import_deduplicator.rs to include Path and is_safe_stdlib_module
- Update call site to use import_deduplicator::collect_imports_from_module()

This is a pure refactoring with no functional changes.
